+++ src/store/main.ts
@@ -1,50 +1,64 @@
 /*
  * @Author       : 卢瑶
  * @Date         : 2022-03-17 17:34:34
- * @LastEditTime : 2022-06-23 14:18:59
+ * @LastEditTime : 2022-06-27 16:42:32
  * @LastEditors  : 卢瑶
  * @Description  : 全局状态管理
  * @FilePath     : /vvt-admin/src/store/main.ts
  */
 
-import { defineStore } from 'pinia'
+import { ref } from 'vue'
+import { defineStore, storeToRefs } from 'pinia'
 import requireAllRoutes from '../router/require-routes'
-import type { IMainActions, IMainState, Theme } from './main-type'
+import type { Theme } from './main-type'
 import router from '../router'
 import { GLOBAL_VARIABLE_NAME } from '../setting/variable-setting'
 import useStorage from '@/hooks/storage'
+import { RouteRecordRaw } from 'vue-router'
 const storage = useStorage()
 
-const useMainStore = defineStore<string, IMainState, {}, IMainActions>('main', {
-  state: () => ({
-    username: storage.getItem(GLOBAL_VARIABLE_NAME.USERNAME)! || '',
-    routes: [],
-    token: storage.getItem(GLOBAL_VARIABLE_NAME.TOKEN)! || '',
-    theme: (storage.getItem(GLOBAL_VARIABLE_NAME.THEME) as Theme) || 'light'
-  }),
-  actions: {
-    registerAsyncRoutes() {
-      const routes = requireAllRoutes()
-      this.routes = routes
-      routes.forEach((route) => {
-        router.addRoute('main', route)
-      })
-    },
-    setName(username: string) {
-      this.username = username
-    },
-    setToken(token: string) {
-      this.token = token
-    },
-    setTheme(theme: Theme) {
-      this.theme = theme
-    }
+const useMainStore = defineStore('main', () => {
+  // state
+  const username = ref<string>(storage.getItem(GLOBAL_VARIABLE_NAME.USERNAME)! || '')
+  const asyncRoutes = ref<RouteRecordRaw[]>([])
+  const token = ref<string>(storage.getItem(GLOBAL_VARIABLE_NAME.TOKEN)! || '')
+  const theme = ref<Theme>((storage.getItem(GLOBAL_VARIABLE_NAME.THEME) as Theme) || 'light')
+
+  // actions
+  const setAsyncRoutes = (newAsyncRoutes: RouteRecordRaw[]) => {
+    asyncRoutes.value = newAsyncRoutes
+  }
+  const setName = (newUsername: string) => {
+    username.value = newUsername
+  }
+  const setToken = (newToken: string) => {
+    token.value = newToken
+  }
+  const setTheme = (newTheme: Theme) => {
+    theme.value = newTheme
+  }
+
+  return {
+    username,
+    asyncRoutes,
+    token,
+    theme,
+
+    setAsyncRoutes,
+    setName,
+    setToken,
+    setTheme
   }
 })
 
 export function setupMainStore() {
   const mainStore = useMainStore()
-  mainStore.registerAsyncRoutes()
+  const { asyncRoutes } = storeToRefs(mainStore)
+  const { setAsyncRoutes } = mainStore
+  setAsyncRoutes(requireAllRoutes())
+  asyncRoutes.value.forEach((route) => {
+    router.addRoute('main', route)
+  })
 }
 
 export default useMainStore